Tests/LibWeb/Text/input/load-event-delayed-by-adopted-style-sheet-background-image.html

<!DOCTYPE html>
<script src="include.js"></script>
<script>
var xhr = new XMLHttpRequest();
xhr.open("POST", "/echo", false);
xhr.setRequestHeader("Content-Type", "application/json");
xhr.send(JSON.stringify({
    method: "GET",
    path: "/delayed-green-adopted.png",
    status: 200,
    headers: {"Content-Type": "image/png", "Cache-Control": "no-store"},
    body: "iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAIAAAD91JpzAAAAD0lEQVR4nGNgaGAAIQgFAA4OAgFhGn2EAAAAAElFTkSuQmCC",
    body_encoding: "base64",
    delay_ms: 500
}));

var sheet = new CSSStyleSheet();
sheet.replaceSync(`#target { background-image: url("/delayed-green-adopted.png"); }`);
document.adoptedStyleSheets = [sheet];
</script>
<div id="target"></div>
<script>
asyncTest((done) => {
    window.addEventListener("load", () => {
        let loadTime = performance.now();
        println(`Load event delayed by at least 400ms: ${loadTime >= 400}`);

        let entries = performance.getEntriesByType("resource");
        let imageEntry = entries.find(e => e.name.includes("delayed-green-adopted"));
        println(`Image resource entry exists: ${imageEntry !== undefined}`);
        println(`Image loaded before load event: ${imageEntry !== undefined && imageEntry.responseEnd <= loadTime}`);
        done();
    });
});
</script>
