Tests/LibWeb/Text/expected/load-event-delayed-by-adopted-style-sheet-background-image.txt

Load event delayed by at least 400ms: true
Image resource entry exists: true
Image loaded before load event: true
